Update int.h to typedef i32 as int for portability

+++ src/fl/int.h
@@ -63,7 +63,7 @@ namespace fl {
         // Default: assume short 16-bit, int 32-bit (uint32_t is unsigned int)
         typedef short i16;
         typedef unsigned short u16;
-        typedef long i32;
+        typedef int i32;             // 'int' is 32-bit on all desktop and most embedded targets
         typedef unsigned int u32;
         typedef long long i64;
         typedef unsigned long long u64;
